- changed some terminating prints

+++ infrastructure-provisioning/aws/src/base/dlab/aws_actions.py
@@ -246,7 +246,7 @@ def remove_ec2(tag_name, tag_value):
                 waiter.wait(InstanceIds=[instance.id])
                 print "The instance " + instance.id + " has been terminated successfully"
         else:
-            print "There are no instances with " + tag_value + " name to terminate"
+            print "There are no instances with '" + tag_name + "' tag to terminate"
     except Exception as err:
         logging.info("Unable to remove EC2: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -346,7 +346,6 @@ def remove_roles_and_profiles(role_name, role_profile_name):
 
 def remove_all_iam_resources(instance_type, scientist=''):
     try:
-        print "[Removing IAM roles, instance profiles and policies]"
         client = boto3.client('iam')
         roles_list = []
         for item in client.list_roles(MaxItems=250).get("Roles"):
@@ -354,7 +353,6 @@ def remove_all_iam_resources(instance_type, scientist=''):
                 roles_list.append(item.get('RoleName'))
         if roles_list:
             roles_list.sort(reverse=True)
-            print roles_list
             for iam_role in roles_list:
                 if '-ssn-Role' in iam_role:
                     if instance_type == 'ssn' or instance_type == 'all':
@@ -362,7 +360,7 @@ def remove_all_iam_resources(instance_type, scientist=''):
                         try:
                             client.delete_role_policy(RoleName=iam_role, PolicyName=os.environ['conf_service_base_name'] + '-ssn-Policy')
                         except:
-                            print 'There is no policy: ' + os.environ['conf_service_base_name'] + '-ssn-Policy'
+                            print 'There is no policy ' + os.environ['conf_service_base_name'] + '-ssn-Policy to delete'
                         remove_roles_and_profiles(iam_role, role_profile_name)
                 if '-edge-Role' in iam_role:
                     if instance_type == 'edge' and scientist in iam_role:
@@ -387,7 +385,7 @@ def remove_all_iam_resources(instance_type, scientist=''):
                             role_profile_name = i.get('InstanceProfileName')
                             remove_roles_and_profiles(iam_role, role_profile_name)
         else:
-            print "There is no IAM role to delete"
+            print "There are no IAM roles instance profiles and policies to delete"
     except Exception as err:
         logging.info("Unable to remove some of the IAM resources: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -415,7 +413,6 @@ def s3_cleanup(bucket, cluster_name, user_name):
 
 def remove_s3(bucket_type='all', scientist=''):
     try:
-        print "[Removing S3 buckets]"
         client = boto3.client('s3')
         bucket_list = []
         if bucket_type == 'ssn':
@@ -433,7 +430,6 @@ def remove_s3(bucket_type='all', scientist=''):
             if list_obj is not None:
                 for o in list_obj:
                     list_obj = o.get('Key')
-                    print list_obj
                     client.delete_objects(
                         Bucket=s3bucket,
                         Delete={'Objects': [{'Key': list_obj}]}
@@ -453,16 +449,17 @@ def remove_s3(bucket_type='all', scientist=''):
 
 def remove_subnets(tag_value):
     try:
-        print "[Removing subnets]"
         ec2 = boto3.resource('ec2')
         client = boto3.client('ec2')
         tag_name = os.environ['conf_service_base_name'] + '-Tag'
         subnets = ec2.subnets.filter(
             Filters=[{'Name': 'tag:{}'.format(tag_name), 'Values': [tag_value]}])
-        for subnet in subnets:
-            print subnet.id
-            client.delete_subnet(SubnetId=subnet.id)
-            print "The subnet " + subnet.id + " has been deleted successfully"
+        if subnets:
+            for subnet in subnets:
+                client.delete_subnet(SubnetId=subnet.id)
+                print "The subnet " + subnet.id + " has been deleted successfully"
+        else:
+            print "There are no private subnets to delete"
     except Exception as err:
         logging.info("Unable to remove subnet: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -474,16 +471,17 @@ def remove_subnets(tag_value):
 
 def remove_sgroups(tag_value):
     try:
-        print "[Removing security groups]"
         ec2 = boto3.resource('ec2')
         client = boto3.client('ec2')
         tag_name = os.environ['conf_service_base_name']
         sgs = ec2.security_groups.filter(
             Filters=[{'Name': 'tag:{}'.format(tag_name), 'Values': [tag_value]}])
-        for sg in sgs:
-            print sg.id
-            client.delete_security_group(GroupId=sg.id)
-            print "The security group " + sg.id + " has been deleted successfully"
+        if sgs:
+            for sg in sgs:
+                client.delete_security_group(GroupId=sg.id)
+                print "The security group " + sg.id + " has been deleted successfully"
+        else:
+            print "There are no security groups to delete"
     except Exception as err:
         logging.info("Unable to remove SG: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -495,14 +493,16 @@ def remove_sgroups(tag_value):
 
 def deregister_image(scientist):
     try:
-        print "[De-registering images]"
         client = boto3.client('ec2')
         response = client.describe_images(
-            Filters=[{'Name': 'name', 'Values': ['{}-{}-notebook-image'.format(os.environ['conf_service_base_name'], scientist)]}])
+            Filters=[{'Name': 'name', 'Values': ['{}-{}'.format(os.environ['conf_service_base_name'], scientist)]}])
         images_list = response.get('Images')
-        for i in images_list:
-            client.deregister_image(ImageId=i.get('ImageId'))
-            print "Notebook AMI has been deregistered successfully"
+        if images_list:
+            for i in images_list:
+                client.deregister_image(ImageId=i.get('ImageId'))
+                print "Notebook AMI " + i.get('ImageId') + " has been deregistered successfully"
+        else:
+            print "There is no notebook ami to deregister"
     except Exception as err:
         logging.info("Unable to de-register image: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -565,9 +565,9 @@ def remove_route_tables(tag_name):
             if rtable:
                 rtable = rtable.get('RouteTableId')
                 client.delete_route_table(RouteTableId=rtable)
-                print "Route table " + rtable + " was removed"
+                print "Route table " + rtable + " has been removed"
             else:
-                print "There is no route table to remove"
+                print "There are no route tables to remove"
     except Exception as err:
         logging.info("Unable to remove route table: " + str(err) + "\n Traceback: " + traceback.print_exc(
             file=sys.stdout))
